Copying prototype1 branch of Project-MurderDome repo to start this project.
There is a lot of code that I could possibly use from Project-MurderDome, but since this is a very different game I will be making an entirely different game a branch did not seem approriate. Also renaming a visual studio solution is not easy, so I just copied everything over. (Which is basically how you rename a visual studio solution)